separate test and field module
