+++ src/mul-float/fma-js.ts
@@ -10,21 +10,21 @@
  * https://github.com/yrrid/submission-wasm-twisted-edwards (see FP51.java and FieldPair.c)
  */
 import { f64, f64x2, func, Module } from "wasmati";
-import { randomGenerators } from "../bigint/field-random.js";
-import { assertDeepEqual } from "../testing/nested.js";
 import { pallasParams } from "../concrete/pasta.params.js";
 import { createField, inverse } from "../bigint/field.js";
 import { assert, bigintFromLimbs, bigintToLimbsRelaxed } from "../util.js";
-import { equivalent, spec, Spec } from "../testing/equivalent.js";
-import { Random } from "../testing/random.js";
 
 export {
   numberToBigint64,
   bigint64ToNumber,
   float52ToInt64,
   int64ToFloat52,
+  madd,
   bigintToFloat51Limbs,
   bigintFromFloat51Limbs,
+  montmulSimple,
+  montmulRef,
+  montmul,
   montmulFma,
   montMulFmaWrapped,
 };
@@ -40,7 +40,7 @@ const maddWasm = func({ in: [f64, f64, f64], out: [f64] }, ([x, y, z]) => {
 });
 let module = Module({ exports: { madd: maddWasm } });
 let { instance } = await module.instantiate();
-let madd = instance.exports.madd;
+let madd: (x: number, y: number, z: number) => number = instance.exports.madd;
 
 let mBytes = new Uint8Array(8);
 let mView = new DataView(mBytes.buffer);
@@ -59,7 +59,6 @@ let c52 = 2 ** 52;
 let c51 = 2 ** 52;
 let c51x3 = 3 * 2 ** 51;
 let c2 = c103 + c51x3;
-let shift51n = 1n << 51n;
 
 // constants we have to subtract after reinterpreting raw float bytes as int64
 let hiPre = numberToBigint64(c103);
@@ -69,33 +68,7 @@ let c51n = numberToBigint64(c51);
 
 // or just mask out everything but the 52 mantissa bits
 let mask51 = (1n << 51n) - 1n;
-let mask52 = (1n << 52n) - 1n;
 let mask64 = (1n << 64n) - 1n;
-
-// random numbers that can be 5 limbs of a number q < 2^255 + 2^253
-// e.g., q < 2p where p is one of the Pasta primes or any < 254 bit prime
-let rng51 = randomGenerators(1n << 51n);
-let rng52 = randomGenerators(1n << 52n);
-let rng51eps = randomGenerators((1n << 51n) + (1n << 49n)); // also works when both x, y are sampled from this
-
-for (let i = 0; i < 10_000; i++) {
-  let x = rng52.randomField(); // note: x can be 52 bits long
-  let y = rng51.randomField();
-  let xF = bigint64ToNumber(x | c52n) - c52;
-  let yF = bigint64ToNumber(y | c52n) - c52;
-
-  let hi = madd(xF, yF, c103);
-  let lo = madd(xF, yF, c2 - hi);
-
-  let loRaw = numberToBigint64(lo);
-  let hiRaw = numberToBigint64(hi);
-
-  let xyBig = x * y;
-  let xyFma = ((hiRaw - hiPre) << 51n) + (loRaw - loPre);
-  let xyFma2 = ((hiRaw & mask52) << 51n) + (loRaw & mask52) - shift51n;
-  assertDeepEqual(xyBig, xyFma);
-  assertDeepEqual(xyBig, xyFma2);
-}
 
 // modmul with 5 x 51-bit limbs
 let Fp = createField(pallasParams.modulus);
@@ -108,9 +81,6 @@ let Rinv = Fp.inverse(R);
 function split(x: bigint) {
   return [x & mask51, x >> 51n];
 }
-
-let field = Spec.field(p, { relaxed: true });
-let fieldStrict = Spec.field(p, { relaxed: false });
 
 // simple, inefficient reference implementation of montgomery multiplication
 
@@ -126,13 +96,6 @@ function montmulRef(xR: bigint, yR: bigint) {
   let zR2 = Fp.multiply(xR, yR);
   return Fp.multiply(zR2, Rinv);
 }
-
-// montmul with the correction step is (exactly) the same as multiplying and dividing by the Montgomery radius
-equivalent({ from: [field, field], to: fieldStrict, verbose: true })(
-  montmulSimple,
-  montmulRef,
-  "montmul ref"
-);
 
 function montmul(x: bigint, y: bigint) {
   let X = bigintToLimbsRelaxed(x, 51, 5);
@@ -166,41 +129,12 @@ function montmul(x: bigint, y: bigint) {
   return bigintFromLimbs(Z, 51, 5);
 }
 
-// the bigint and array-based versions of montmul are equivalent
-equivalent({ from: [field, field], to: field, verbose: true })(
-  montmulSimple,
-  montmul,
-  "montmul consistent"
-);
-
-// given inputs < p, montmul returns a value < 2p
-// this means it will need a correction step in many places! (we can't go from 2p to 2p, because the Montgomery radius is too small)
-equivalent({ from: [field, field], to: Spec.boolean, verbose: true })(
-  (x, y) => montmul(x, y) < 2n * p,
-  () => true,
-  "montmul < 2p"
-);
-
 function int64ToFloat52(x: bigint) {
   return bigint64ToNumber(x | c51n) - c51;
 }
 function float52ToInt64(x: number) {
   return numberToBigint64(x + c51) & mask51;
 }
-
-let int51 = spec(Random(() => rng51.randomField()));
-let int51F = spec(Random(() => Number(rng51.randomField())));
-
-equivalent({ from: [int51], to: int51, verbose: true })(
-  (x) => x,
-  (x) => float52ToInt64(int64ToFloat52(x)),
-  "int/float roundtrip"
-);
-equivalent({ from: [int51F], to: int51F, verbose: true })(
-  (x) => x,
-  (x) => int64ToFloat52(float52ToInt64(x)),
-  "float/int roundtrip"
-);
 
 // store a limb vector of float64s
 function bigintToFloat51Limbs(x: bigint) {
@@ -287,10 +221,3 @@ function montMulFmaWrapped(x: bigint, y: bigint) {
   let Z = montmulFma(X, Y);
   return bigintFromFloat51Limbs(Z);
 }
-
-// montmulFma is exactly equivalent to montmul
-equivalent({ from: [field, field], to: fieldStrict, verbose: true })(
-  montmul,
-  montMulFmaWrapped,
-  "montmul fma (js)"
-);
